requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Security
argon2-cffi==23.1.0

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# File Processing
Pillow==10.1.0
PyPDF2==3.0.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Google Calendar
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utilities
python-dotenv==1.0.0

# Production
gunicorn==21.2.0

